project/requirements.txt
streamlit
opencv-python-headless
numpy
torch
torchvision
diskcache
pytest
scikit-learn
Pillow
